Ignore future jobs in `jobs:check`.
If there are scheduled jobs in the future, they should not be included in the jobs:check, which is intended to look for jobs that should be processed but have not yet been processed.

To accomplish this we check if the `run_at` is before the `max_age` (default: 300 seconds) instead of the `created_at`.